monolith/requirements.txt
Django==5.2.10
pymysql==1.1.0
djangorestframework==3.15.2
django-filter==24.3
